feat: 이미지 요소 바운드 조정 &&  변환 행렬 추출
